+++ go.py
@@ -2,6 +2,7 @@
 Cecil, it all starts here.
 '''
 import json
+from os import listdir
 from datetime import datetime, timedelta
 from typing import List
 from pathlib import Path
@@ -12,14 +13,11 @@ from passlib.context import CryptContext
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy import create_engine
 
-from os import listdir
-from pathlib import Path
 from baquet.user import Directory, User
 from baquet.watchlist import Watchlist
 
-import sql_models
-import models
-import control
+import orm_models
+import json_models
 from constants import CecilConstants
 
 
@@ -36,8 +34,14 @@ def _exists(directoryname, filename):
     Test existance of a file.
     '''
 
-    if not Path(f"./{directoryname}/{filename}.db").exists():
-        raise FileNotFoundError
+    try:
+        if not Path(f"./{directoryname}/{filename}.db").exists():
+            raise FileNotFoundError
+    except FileExistsError:
+        raise HTTPException(
+            status_code=404,
+            detail=f'{directoryname[0].upper() + directoryname[1:-1]}: {filename}, does not exist.'
+        )
 
 
 def _user_helper(user_id):
@@ -56,11 +60,11 @@ def _make_config():
 
 
 def _init_roles(session):
-    deactivated = sql_models.Role(
+    deactivated = orm_models.Role(
         role_id=CecilConstants.DEACTIVATED_ROLE, role_name="deactivated")
-    admin = sql_models.Role(
+    admin = orm_models.Role(
         role_id=CecilConstants.ADMIN_ROLE, role_name="admin")
-    non_privileged = sql_models.Role(
+    non_privileged = orm_models.Role(
         role_id=CecilConstants.NON_PRIVILEGED_ROLE, role_name="non-privileged")
 
     roles = [deactivated, admin, non_privileged]
@@ -70,7 +74,7 @@ def _init_roles(session):
 
 
 def _init_admin(session):
-    admin = sql_models.User(
+    admin = orm_models.User(
         username="admin",
         hashed_password=PWD_CONTEXT.hash("password"),
         role=CecilConstants.ADMIN_ROLE,
@@ -83,7 +87,7 @@ def _init_admin(session):
 
 def _init_db(database, engine, session):
     database.parent.mkdir(parents=True, exist_ok=True)
-    sql_models.BASE.metadata.create_all(engine)
+    orm_models.BASE.metadata.create_all(engine)
 
     _init_roles(session)
     _init_admin(session)
@@ -120,7 +124,7 @@ def get_authuser(username: str):
     '''
     Get the user from the cecil DB for authentication.
     '''
-    return CONN.query(sql_models.User).filter(sql_models.User.username == username).first()
+    return CONN.query(orm_models.User).filter(orm_models.User.username == username).first()
 
 
 def authenticate_user(username: str, password: str):
@@ -169,7 +173,7 @@ async def get_current_user(token: str = Depends(OAUTH2_SCHEME)):
         username: str = payload.get("sub")
         if username is None:
             raise credentials_exception
-        token_data = models.TokenData(username=username)
+        token_data = json_models.TokenData(username=username)
     except JWTError:
         raise credentials_exception
     user = get_authuser(username=token_data.username)
@@ -178,7 +182,7 @@ async def get_current_user(token: str = Depends(OAUTH2_SCHEME)):
     return user
 
 
-async def get_current_active_user(current_user: models.AuthUser = Depends(get_current_user)):
+async def get_current_active_user(current_user: json_models.AuthUser = Depends(get_current_user)):
     '''
     Get the current user and ensure they are active.
     '''
@@ -187,7 +191,9 @@ async def get_current_active_user(current_user: models.AuthUser = Depends(get_cu
     return current_user
 
 
-async def get_current_admin_user(current_user: models.AuthUser = Depends(get_current_active_user)):
+async def get_current_admin_user(
+        current_user: json_models.AuthUser = Depends(get_current_active_user)
+):
     '''
     Get the current user, ensure they are active and an admin.
     '''
@@ -201,22 +207,22 @@ CONN = _make_conn()
 # CECIL USER OPERATIONS
 
 
-@CECIL.get("/admin/invite_codes/", response_model=List[models.InviteCode])
+@CECIL.get("/admin/invite_codes/", response_model=List[json_models.InviteCode])
 async def get_invite_codes(
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_admin_user
         )
 ):
     '''
     List the created invite codes.
     '''
-    return CONN.query(sql_models.InviteCode).all()
+    return CONN.query(orm_models.InviteCode).all()
 
 
 @CECIL.post("/admin/invite_codes/")
 async def post_invite_code(
-        invite_c: models.AddText,
-        current_user: models.AuthUser = Depends(get_current_admin_user)
+        invite_c: json_models.AddText,
+        current_user: json_models.AuthUser = Depends(get_current_admin_user)
 ):
     '''
     Generate an invite code to create new users.
@@ -226,7 +232,7 @@ async def post_invite_code(
     expires_at = created_at + \
         timedelta(minutes=CONFIG.get(
             CecilConstants.ACCESS_TOKEN_EXPIRE_MINUTES))
-    invite = sql_models.InviteCode(
+    invite = orm_models.InviteCode(
         hashed_invite_code=invite_code_hash,
         created_at=created_at,
         created_by=current_user.user_id,
@@ -236,7 +242,7 @@ async def post_invite_code(
     CONN.commit()
 
 
-@CECIL.get("/admin/users/", response_model=List[models.AuthUser])
+@CECIL.get("/admin/users/", response_model=List[json_models.AuthUser])
 async def get_authusers(
         current_user=Depends(  # pylint: disable=unused-argument
             get_current_admin_user)
@@ -244,7 +250,7 @@ async def get_authusers(
     '''
     Get all cecil users.
     '''
-    return CONN.query(sql_models.User).all()
+    return CONN.query(orm_models.User).all()
 
 
 @CECIL.post("/admin/users/{user_id}/deactivate")
@@ -256,8 +262,8 @@ async def deactivate_user(
     '''
     Deactivate a specific user.
     '''
-    user = CONN.query(sql_models.User).filter(
-        sql_models.User.user_id == user_id)
+    user = CONN.query(orm_models.User).filter(
+        orm_models.User.user_id == user_id)
     user.role = CecilConstants.DEACTIVATED_ROLE
     CONN.commit()
 
@@ -265,26 +271,26 @@ async def deactivate_user(
 @CECIL.delete("/admin/invite_codes/{invite_code_id}")
 async def delete_invite_code(
         invite_code_id: int,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_admin_user
         )
 ):
     '''
     Delete an invite code manually.
     '''
-    invite_code = CONN.query(sql_models.InviteCode).filter(
-        sql_models.InviteCode.invite_id == invite_code_id
+    invite_code = CONN.query(orm_models.InviteCode).filter(
+        orm_models.InviteCode.invite_id == invite_code_id
     ).first()
     CONN.delete(invite_code)
     CONN.commit()
 
 
 @CECIL.post("/register")
-async def register(registration_data: models.RegistrationData):
+async def register(registration_data: json_models.RegistrationData):
     '''
     Register with an invite code.
     '''
-    invite_codes = CONN.query(sql_models.InviteCode).all()
+    invite_codes = CONN.query(orm_models.InviteCode).all()
     invited = False
     for invite in invite_codes:
         if verify_password(registration_data.invite_code, invite.hashed_invite_code):
@@ -292,7 +298,7 @@ async def register(registration_data: models.RegistrationData):
             chosen = invite
 
     if invited:
-        new_user = sql_models.User(
+        new_user = orm_models.User(
             username=registration_data.username,
             hashed_password=get_password_hash(registration_data.password),
             role=CecilConstants.NON_PRIVILEGED_ROLE,
@@ -304,7 +310,7 @@ async def register(registration_data: models.RegistrationData):
         CONN.commit()
 
 
-@CECIL.post("/token", response_model=models.Token)
+@CECIL.post("/token", response_model=json_models.Token)
 async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
     '''
     Get a token for access after logging in.
@@ -326,8 +332,8 @@ async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends(
 
 @CECIL.post("/update_password/")
 async def update_password(
-        update_password_request: models.UpdatePassword,
-        current_user: models.AuthUser = Depends(
+        update_password_request: json_models.UpdatePassword,
+        current_user: json_models.AuthUser = Depends(
             get_current_active_user
         )
 ):
@@ -353,11 +359,11 @@ async def update_password(
 # TWITTER OPERATIONS
 
 
-@CECIL.get("/users/", response_model=models.PaginateUser)
+@CECIL.get("/users/", response_model=json_models.PaginateUser)
 async def get_users(
         page: int = 1,
         page_size: int = 20,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
@@ -370,8 +376,8 @@ async def get_users(
 
 @CECIL.post("/users/")
 async def add_user(
-        user: models.AddUser,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        user: json_models.AddUser,
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
@@ -381,105 +387,107 @@ async def add_user(
     User(user.user_id).get_user()
 
 
-@CECIL.get("/users/{user_id}", response_model=models.User)
+@CECIL.get("/users/{user_id}", response_model=json_models.User)
 async def get_user(
         user_id: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get a user's top level info.
     '''
-    try:
-        user = _user_helper(user_id)
-    except FileNotFoundError:
-        raise HTTPException(
-            status_code=404, detail=f'User, {user_id}, does not exist.')
-
+    user = _user_helper(user_id)
     return user.get_user()
 
 
-@CECIL.get("/users/{user_id}/favorites/", response_model=models.PaginateFavorites)
+@CECIL.get("/users/{user_id}/favorites/", response_model=json_models.PaginateFavorites)
 async def get_favorites(
         user_id: str,
         page: int = 1,
         page_size: int = 20,
         watchlist_id: str = None,
         watchwords_id: str = None,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get a user's favorites.
     '''
-    try:
-        response = control.get_favorites(
-            user_id, page, page_size, watchlist_id, watchwords_id)
-    except FileNotFoundError:
-        raise HTTPException(
-            status_code=404, detail=f'User, {user_id}, does not exist.')
+    user = _user_helper(user_id)
+    if watchlist_id:
+        watchlist_id = _wl_helper(watchlist_id)
+    if watchwords_id:
+        watchwords_id = _wl_helper(watchwords_id)
+    return user.get_favorites(
+        page=page,
+        page_size=page_size,
+        watchlist=watchlist_id,
+        watchwords=watchwords_id
+    )
 
-    return response
 
-
-@CECIL.get("/users/{user_id}/favorites/tags", response_model=models.PaginateFavorites)
+@CECIL.get("/users/{user_id}/favorites/tags", response_model=json_models.PaginateFavorites)
 async def get_tags_favorites(
         user_id: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get a list of tags that apply to one or more of a user's favorites.
     '''
-    return control.get_tags_favorites(user_id)
+    user = _user_helper(user_id)
+    return user.get_tags("favorite")
 
 
-@CECIL.get("/users/{user_id}/favorites/tags/{tag_id}", response_model=models.PaginateFavorites)
+@CECIL.get("/users/{user_id}/favorites/tags/{tag_id}", response_model=json_models.PaginateFavorites)
 async def get_favorites_tagged(
         user_id: str,
         tag_id: int,
         page: int = 1,
         page_size: int = 20,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get a list of favorites that have this particular tag.
     '''
-    return control.get_favorites_tagged(user_id, tag_id, page, page_size)
+    user = _user_helper(user_id)
+    return user.get_favorites_tagged(tag_id, page, page_size=page_size)
 
 
 @CECIL.get("/users/{user_id}/favorite/{tweet_id}/notes/")
 async def get_notes_favorite(
         user_id: str,
         tweet_id: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get a list of notes on a particular tweet.
     '''
-    return control.get_notes_favorite(user_id, tweet_id)
+    user = _user_helper(user_id)
+    return user.get_notes_favorite(tweet_id)
 
 
 @CECIL.post("/users/{user_id}/favorite/{tweet_id}/notes/")
 async def add_note_favorite(
         user_id: str,
         tweet_id: str,
-        note: models.AddText,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        note: json_models.AddText,
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Add a note to a user's timeline tweet.
     '''
-    control.add_note_favorite(user_id, tweet_id, note.text)
+    user = _user_helper(user_id)
+    user.add_note_favorite(tweet_id, note.text)
 
 
 @CECIL.delete("/users/{user_id}/favorite/{tweet_id}/notes/{note_id}")
@@ -487,43 +495,46 @@ async def remove_note_favorite(
         user_id: str,
         tweet_id: str,
         note_id: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Delete a note from user's timeline tweet.
     '''
-    control.remove_note_favorite(user_id, tweet_id, note_id)
+    user = _user_helper(user_id)
+    user.remove_note_favorite(tweet_id, note_id)
 
 
-@CECIL.get("/users/{user_id}/favorites/{tweet_id}/tags/", response_model=List[models.Tag])
+@CECIL.get("/users/{user_id}/favorites/{tweet_id}/tags/", response_model=List[json_models.Tag])
 async def get_tags_favorite(
         user_id: str,
         tweet_id: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get the tags on a particular tweet.
     '''
-    return control.get_tags_favorite(user_id, tweet_id)
+    user = _user_helper(user_id)
+    return user.get_tags_favorite(tweet_id)
 
 
 @CECIL.post("/users/{user_id}/favorites/{tweet_id}/tags/")
 async def add_tag_favorite(
         user_id: str,
         tweet_id: str,
-        tag: models.AddText,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        tag: json_models.AddText,
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Add a tag to a user's timeline tweet.
     '''
-    control.add_tag_favorite(user_id, tweet_id, tag.text)
+    user = _user_helper(user_id)
+    user.add_tag_favorite(tweet_id, tag.text)
 
 
 @CECIL.delete("/users/{user_id}/favorites/{tweet_id}/tags/{tag_id}")
@@ -531,154 +542,166 @@ async def remove_tag_favorite(
         user_id: str,
         tweet_id: str,
         tag_id: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Delete a tag from user's timeline tweet.
     '''
-    control.remove_tag_favorite(user_id, tweet_id, tag_id)
+    user = _user_helper(user_id)
+    user.remove_tag_timeline(tweet_id, tag_id)
 
 
-@CECIL.get("/users/{user_id}/followers/", response_model=models.PaginateFriendsOrFollowing)
+@CECIL.get("/users/{user_id}/followers/", response_model=json_models.PaginateFriendsOrFollowing)
 async def get_followers(
         user_id: str,
         page: int = 1,
         page_size: int = 1500,
         watchlist_id: str = None,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get a user's followers.
     '''
-    return control.get_followers(user_id, page, page_size, watchlist_id)
+    user = _user_helper(user_id)
+    if watchlist_id:
+        watchlist_id = _wl_helper(watchlist_id)
+
+    return user.get_followers(
+        page=page, page_size=page_size, watchlist=watchlist_id)
 
 
-@CECIL.get("/users/{user_id}/friends/", response_model=models.PaginateFriendsOrFollowing)
+@CECIL.get("/users/{user_id}/friends/", response_model=json_models.PaginateFriendsOrFollowing)
 async def get_friends(
         user_id: str,
         page: int = 1,
         page_size: int = 1500,
         watchlist_id: str = None,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get a user's friends.
     '''
-    try:
-        response = control.get_friends(
-            user_id, page, page_size, watchlist_id
-        )
-    except FileNotFoundError:
-        raise HTTPException(
-            status_code=404, detail=f'User, {user_id}, does not exist.')
+    user = _user_helper(user_id)
+    if watchlist_id:
+        watchlist_id = _wl_helper(watchlist_id)
 
-    return response
+    return user.get_friends(
+        page=page, page_size=page_size, watchlist=watchlist_id)
 
 
-@CECIL.get("/users/{user_id}/notes/", response_model=models.PaginateUserNotes)
+@CECIL.get("/users/{user_id}/notes/", response_model=json_models.PaginateUserNotes)
 async def get_notes_user(
         user_id: str,
         page: int = 1,
         page_size: int = 20,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get the notes about a user.
     '''
-    return control.get_notes_user(user_id, page, page_size)
+    user = _user_helper(user_id)
+    return user.get_notes_user(page=page, page_size=page_size)
 
 
 @CECIL.post("/users/{user_id}/notes/")
 async def add_note_user(
         user_id: str,
-        note: models.AddText,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        note: json_models.AddText,
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Add a note to a user's file.
     '''
-    control.add_note_user(user_id, note.text)
+    user = _user_helper(user_id)
+    user.add_note_user(note.text)
 
 
 @CECIL.delete("/users/{user_id}/notes/{note_id}")
 async def remove_note_user(
         user_id: str,
         note_id: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Remove a note from a user's file.
     '''
-    control.remove_note_user(user_id, note_id)
+    user = _user_helper(user_id)
+    user.remove_note_user(note_id)
 
 
-@CECIL.get("/users/{user_id}/stats/{watchlist_id}", response_model=models.UserStats)
+@CECIL.get("/users/{user_id}/stats/{watchlist_id}", response_model=json_models.UserStats)
 async def get_stats(
         user_id: str,
         watchlist_id: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get a user's friends.
     '''
-    try:
-        response = control.get_stats(user_id, watchlist_id)
-    except FileNotFoundError:
-        raise HTTPException(
-            status_code=404, detail=f'User, {user_id}, does not exist.')
+    user = _user_helper(user_id)
+    watchlist = _wl_helper(watchlist_id)
+    return {
+        'followers_watchlist_percent': user.get_followers_watchlist_percent(watchlist=watchlist),
+        'followers_watchlist_completion': user.get_followers_watchlist_completion(
+            watchlist=watchlist
+        ),
+        'friends_watchlist_percent': user.get_friends_watchlist_percent(watchlist=watchlist),
+        'friends_watchlist_completion': user.get_friends_watchlist_completion(watchlist=watchlist),
+        'favorite_watchlist_percent': user.get_favorite_watchlist_percent(watchlist=watchlist),
+        'retweet_watchlist_percent': user.get_retweet_watchlist_percent(watchlist=watchlist)
+    }
 
-    return response
 
-
-@CECIL.get("/users/{user_id}/timeline/", response_model=models.PaginateTimeline)
+@CECIL.get("/users/{user_id}/timeline/", response_model=json_models.PaginateTimeline)
 async def get_timeline(
         user_id: str,
         page: int = 1,
         page_size: int = 20,
         watchlist_id: str = None,
         watchwords_id: str = None,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get a user's timeline.
     '''
-    try:
-        response = control.get_timeline(
-            user_id, page, page_size, watchlist_id, watchwords_id)
-    except FileNotFoundError:
-        raise HTTPException(
-            status_code=404, detail=f'User, {user_id}, does not exist.')
-
-    return response
+    user = _user_helper(user_id)
+    if watchlist_id:
+        watchlist_id = _wl_helper(watchlist_id)
+    if watchwords_id:
+        watchwords_id = _wl_helper(watchwords_id)
+    return user.get_timeline(
+        page, page_size, watchlist=watchlist_id, watchwords=watchwords_id
+    )
 
 
 @CECIL.get("/users/{user_id}/timeline/tags/")
 async def get_tags_timelines(
         user_id: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get a timeline's tags.
     '''
-    return control.get_tags_timelines(user_id)
+    user = _user_helper(user_id)
+    return user.get_tags("timeline")
 
 
 @CECIL.get("/users/{user_id}/timeline/tags/{tag_id}")
@@ -687,43 +710,46 @@ async def get_timeline_tagged(
         tag_id: str,
         page: int = 1,
         page_size: int = 20,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get timelines tagged.
     '''
-    return control.get_timeline_tagged(user_id, tag_id, page, page_size)
+    user = _user_helper(user_id)
+    return user.get_timeline_tagged(tag_id, page, page_size)
 
 
 @CECIL.get("/users/{user_id}/timeline/{tweet_id}/notes/")
 async def get_notes_timeline(
         user_id: str,
         tweet_id: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get timeline tweet notes.
     '''
-    return control.get_notes_timeline(user_id, tweet_id)
+    user = _user_helper(user_id)
+    return user.get_notes_timeline(tweet_id)
 
 
 @CECIL.post("/users/{user_id}/timeline/{tweet_id}/notes/")
 async def add_note_timeline(
         user_id: str,
         tweet_id: str,
-        note: models.AddText,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        note: json_models.AddText,
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Add a note to a user's timeline tweet.
     '''
-    control.add_note_timeline(user_id, tweet_id, note.text)
+    user = _user_helper(user_id)
+    user.add_note_timeline(tweet_id, note.text)
 
 
 @CECIL.delete("/users/{user_id}/timeline/{tweet_id}/notes/{note_id}")
@@ -731,43 +757,46 @@ async def remove_note_timeline(
         user_id: str,
         tweet_id: str,
         note_id: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Delete a note from user's timeline tweet.
     '''
-    control.remove_note_timeline(user_id, tweet_id, note_id)
+    user = _user_helper(user_id)
+    user.remove_note_timeline(tweet_id, note_id)
 
 
-@CECIL.get("/users/{user_id}/timeline/{tweet_id}/tags/", response_model=List[models.Tag])
+@CECIL.get("/users/{user_id}/timeline/{tweet_id}/tags/", response_model=List[json_models.Tag])
 async def get_tags_timeline(
         user_id: str,
         tweet_id: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get the tags on a timeline tweet.
     '''
-    return control.get_tags_timeline(user_id, tweet_id)
+    user = _user_helper(user_id)
+    return user.get_tags_timeline(tweet_id)
 
 
 @CECIL.post("/users/{user_id}/timeline/{tweet_id}/tags/")
 async def add_tag_timeline(
         user_id: str,
         tweet_id: str,
-        tag: models.AddText,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        tag: json_models.AddText,
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Add a tag to a user's timeline tweet.
     '''
-    control.add_tag_timeline(user_id, tweet_id, tag.text)
+    user = _user_helper(user_id)
+    user.add_tag_timeline(tweet_id, tag.text)
 
 
 @CECIL.delete("/users/{user_id}/timeline/{tweet_id}/tags/{tag_id}")
@@ -775,131 +804,122 @@ async def remove_tag_timeline(
         user_id: str,
         tweet_id: str,
         tag_id: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Delete a tag from user's timeline tweet.
     '''
-    control.remove_tag_timeline(user_id, tweet_id, tag_id)
+    user = _user_helper(user_id)
+    user.remove_tag_timeline(tweet_id, tag_id)
 
 
 @CECIL.get("/watchlists/", response_model=List[str])
 async def get_watchlists(
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get a list of watchlists in the watchlist directory.
     '''
-    return control.get_watchlists()
+    if _WL_PATH.exists():
+        result = [fn.split(".")[0] for fn in listdir(_WL_PATH)]
+    else:
+        result = []
+
+    return result
 
 
 @CECIL.post("/watchlists/")
 async def add_watchlist(
-        watchlist: models.AddWatchlist,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        watchlist: json_models.AddWatchlist,
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Create a watchlist.
     '''
-    return control.create_watchlist(watchlist.watchlist_id)
+    Watchlist(watchlist.watchlist_id)
 
 
-@CECIL.get("/watchlists/{watchlist_name}", response_model=models.WatchlistInfo)
+@CECIL.get("/watchlists/{watchlist_id}", response_model=json_models.WatchlistInfo)
 async def get_watchlist(
-        watchlist_name: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        watchlist_id: str,
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get top level details of a watchlist.
     '''
-    try:
-        response = control.get_watchlist(watchlist_name)
-    except FileNotFoundError:
-        raise HTTPException(
-            status_code=404, detail=f'Watchlist, {watchlist_name}, does not exist.')
+    watchlist = _wl_helper(watchlist_id)
 
-    return response
+    return {
+        'name': watchlist_id,
+        'watchlist_count': watchlist.get_watchlist_count(),
+        'watchword_count': watchlist.get_watchwords_count()
+    }
 
 
-@CECIL.get("/watchlists/{watchlist_name}/users/", response_model=List[models.User])
+@CECIL.get("/watchlists/{watchlist_id}/users/", response_model=List[json_models.User])
 async def get_watchlist_users(
-        watchlist_name: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        watchlist_id: str,
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get users on the watchlist.
     '''
-    try:
-        response = control.get_watchlist_users(watchlist_name)
-    except FileNotFoundError:
-        raise HTTPException(
-            status_code=404, detail=f'Watchlist, {watchlist_name}, does not exist.')
+    watchlist = _wl_helper(watchlist_id)
+    watchlist.refresh_watchlist_user_data()
 
-    return response
+    return watchlist.get_watchlist_users()
 
 
-@CECIL.post("/watchlists/{watchlist_name}/users/")
+@CECIL.post("/watchlists/{watchlist_id}/users/")
 async def add_watchlist_users(
-        watchlist_name: str,
-        user: models.AddUser,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        watchlist_id: str,
+        user: json_models.AddUser,
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Add user to the watchlist.
     '''
-    try:
-        control.add_watchlist(watchlist_name, user.user_id)
-    except FileNotFoundError:
-        raise HTTPException(
-            status_code=404,
-            detail=f'Watchlist, {watchlist_name}, does not exist.'
-        )
+    watchlist = _wl_helper(watchlist_id)
+    user = _user_helper(user.user_id)
+    watchlist.add_watchlist(user)
 
 
-@CECIL.get("/watchlists/{watchlist_name}/words/", response_model=List[str])
+@CECIL.get("/watchlists/{watchlist_id}/words/", response_model=List[str])
 async def get_watchwords(
-        watchlist_name: str,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        watchlist_id: str,
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Get watchwords.
     '''
-    try:
-        response = control.get_watchwords(watchlist_name)
-    except FileNotFoundError:
-        raise HTTPException(
-            status_code=404, detail=f'Watchlist, {watchlist_name}, does not exist.')
-
-    return response
+    watchlist = _wl_helper(watchlist_id)
+    return watchlist.get_watchwords()
 
 
-@CECIL.post("/watchlists/{watchlist_name}/words/")
+@CECIL.post("/watchlists/{watchlist_id}/words/")
 async def add_watchword(
-        watchlist_name: str,
-        watchword: models.AddText,
-        current_user: models.AuthUser = Depends(  # pylint: disable=unused-argument
+        watchlist_id: str,
+        watchword: json_models.AddText,
+        current_user: json_models.AuthUser = Depends(  # pylint: disable=unused-argument
             get_current_active_user
         )
 ):
     '''
     Add a search term to the watchwords.
     '''
-    try:
-        control.add_watchword(watchlist_name, watchword.text)
-    except FileNotFoundError:
-        raise HTTPException(
-            status_code=404, detail=f'Watchlist, {watchlist_name}, does not exist.')
+    watchlist = _wl_helper(watchlist_id)
+    watchlist.add_watchword(watchword)
